api for register company, insert a new client, get all clients, get one client

+++ spring-angular/src/main/java/com/example/springangularapp/repository/ClientRepository.java
@@ -6,8 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-
-public interface ClientRepository extends CrudRepository<Client, Long> {
+public interface ClientRepository extends CrudRepository<Client, Integer> {
     List<Client> findAll();
 
     Client findByCnp(String email);
